Test for duplicated file copy in PULP_MANIFEST on updated iso copy
ISO repo copy can lead to a duplicates in a target repo and corrupted manifest.

Upload an ISO into a source iso-repo and copy the ISO to the target. Update
the ISO on the source to have different content and copy the updated ISO to
the target iso-repo.

Verify that the target repo only has one entry for the ISO in the
``PULP_MANIFEST``

See https://pulp.plan.io/issues/4943